requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
boto3==1.34.34
botocore==1.34.162
celery==5.3.4
certifi==2026.1.4
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-tailwind==0.5.0
cron_descriptor==2.0.6
cryptography==46.0.4
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.12.0
django-filter==23.5
django-rest-framework==0.1.0
django-restframework==0.0.1
django-storages==1.14.2
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
idna==3.11
jmespath==1.1.0
kombu==5.6.2
msgpack==1.1.2
oauthlib==3.3.1
packaging==26.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pycparser==3.0
PyJWT==2.11.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
redis==5.0.1
requests==2.32.5
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
s3transfer==0.10.4
simplejwt==2.0.1
six==1.17.0
sqlparse==0.5.5
typing==3.7.4.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.3
whitenoise==6.6.0
